requirements.txt
prometheus_client
python-dotenv
httpx
roboflow
opencv-python-headless